# public/openapi/cascade-partner-v1.json
{"openapi":"3.1.0","info":{"title":"Cascade Logistics Partner API","version":"1.0.0"},"servers":[{"url":"https://cascadelogistics.vercel.app/api/v1"}],"security":[{"bearerAuth":[]}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}},"responses":{"ok":{"description":"Successful JSON response"},"error":{"description":"Safe API error"}}},"paths":{"/me":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/uploads":{"post":{"responses":{"201":{"$ref":"#/components/responses/ok"}}}},"/shipments":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}},"post":{"parameters":[{"name":"Idempotency-Key","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"201":{"$ref":"#/components/responses/ok"},"422":{"$ref":"#/components/responses/error"}}}},"/shipments/{shipmentId}":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}},"patch":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/shipments/{shipmentId}/cancel":{"post":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/shipments/{shipmentId}/documents":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/shipments/{shipmentId}/documents/{documentId}":{"get":{"responses":{"200":{"description":"Private file"}}},"delete":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/shipments/{shipmentId}/timeline":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/shipments/{shipmentId}/tracking-numbers":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/shipments/{shipmentId}/invoice":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/shipments/{shipmentId}/invoice/file":{"get":{"responses":{"200":{"description":"Private file"}}}},"/shipments/{shipmentId}/payment-proofs":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}},"post":{"responses":{"201":{"$ref":"#/components/responses/ok"}}}},"/payment-proofs/{paymentProofId}/file":{"get":{"responses":{"200":{"description":"Private file"}}}},"/webhook-endpoints":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}},"post":{"responses":{"201":{"$ref":"#/components/responses/ok"}}}},"/webhook-endpoints/{endpointId}":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}},"patch":{"responses":{"200":{"$ref":"#/components/responses/ok"}}},"delete":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/webhook-endpoints/{endpointId}/test":{"post":{"responses":{"202":{"$ref":"#/components/responses/ok"}}}},"/webhook-deliveries":{"get":{"responses":{"200":{"$ref":"#/components/responses/ok"}}}},"/webhook-deliveries/{deliveryId}/replay":{"post":{"responses":{"202":{"$ref":"#/components/responses/ok"}}}}}}